+++ src/renderer/gl/gl_renderer_impl.hpp
@@ -8,6 +8,7 @@
 #pragma once
 
 #include "vglx/core/renderer.hpp"
+#include "vglx/core/window.hpp"
 #include "vglx/materials/depth_material.hpp"
 #include "vglx/math/matrix4.hpp"
 #include "vglx/math/vector2.hpp"
@@ -62,13 +63,13 @@ public:
     auto operator=(const Impl&) -> Impl& = delete;
     auto operator=(Impl&&) -> Impl& = delete;
 
-    [[nodiscard]] auto Initialize() -> std::expected<void, std::string>;
+    [[nodiscard]] auto Initialize(Window::Impl& window) -> std::expected<void, std::string>;
 
     auto Render(Scene* scene, Camera* camera, RenderTarget* target = nullptr) -> void;
 
     auto Clear(RenderTarget* target = nullptr) -> void;
 
-    auto SetViewport(int x, int y, int width, int height, Vector2 content_scale) -> void;
+    auto SetViewport(const Viewport& viewport) -> void;
 
     auto SetClearColor(const Color& color) -> void;
 
@@ -115,8 +116,10 @@ private:
     CameraUniforms camera_ {};
     GLUniformBuffer camera_uniforms_ {"ub_Camera", sizeof(CameraUniforms)};
 
-    int viewport_width_ {0};
-    int viewport_height_ {0};
+    Window::Impl* window_ {nullptr};
+
+    Viewport viewport_ {};
+    bool viewport_pinned_ {false};
 
     std::unique_ptr<RenderLists> render_lists_;
     std::unique_ptr<RenderLists> shadow_render_lists_;
@@ -160,6 +163,8 @@ private:
     auto UpdateFrameUniforms() -> void;
 
     auto UpdateCameraUniforms(Camera* camera) -> void;
+
+    auto SyncWithWindow() -> void;
 };
 
 }